requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dateparser==1.2.0
Flask==3.0.3
GoogleNews==1.6.14
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
newsapi-python==0.2.7
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
requests==2.32.2
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.0
tzlocal==5.2
urllib3==2.2.1
Werkzeug==3.0.3
